feat(server): enhance link management with improved schemas, repository methods, and service validations

+++ apps/server/src/modules/link/link.service.ts
@@ -3,6 +3,7 @@ import type { LinkRepository } from "./link.repository";
 import type {
 	createLinkType,
 	deleteLinkType,
+	getLinksType,
 	updateLinkType,
 } from "./link.types";
 
@@ -13,45 +14,74 @@ export class LinkService {
 		this.linkRepository = linkRepository;
 	}
 
-	async createLink(data: createLinkType & { userId: string }) {
-		const createdLink = await this.linkRepository.createLink(data);
+	getLinks = async ({ userId }: getLinksType) => {
+		if (!userId) throw new HTTPException(400, { message: "Missing userId" });
+
+		const links = await this.linkRepository.getLinks({ userId });
+
+		if (!links)
+			throw new HTTPException(404, {
+				message: "Links not found or not exists",
+			});
+
+		return links;
+	};
+
+	createLink = async ({ index, title, url, userId }: createLinkType) => {
+		if (!index) throw new HTTPException(400, { message: "Missing index" });
+		if (!title) throw new HTTPException(400, { message: "Missing title" });
+		if (!url) throw new HTTPException(400, { message: "Missing url" });
+		if (!userId) throw new HTTPException(400, { message: "Missing userId" });
+
+		const createdLink = await this.linkRepository.createLink({
+			index,
+			title,
+			url,
+			userId,
+		});
+
+		if (!createdLink)
+			throw new HTTPException(500, {
+				message: "Failed to create link",
+			});
 
 		return createdLink;
-	}
+	};
 
-	async updateLink({ id, title, href }: updateLinkType, userId: string) {
-		const link = await this.linkRepository.getLinkById({ id });
-
-		if (!link) {
-			throw new HTTPException(404, { message: "Link not found" });
-		}
-
-		if (link.userId !== userId) {
-			throw new HTTPException(403, { message: "Forbidden" });
-		}
+	updateLink = async ({ id, userId, index, title, url }: updateLinkType) => {
+		if (!id) throw new HTTPException(400, { message: "Missing id" });
+		if (!index) throw new HTTPException(400, { message: "Missing index" });
+		if (!title) throw new HTTPException(400, { message: "Missing title" });
+		if (!url) throw new HTTPException(400, { message: "Missing url" });
+		if (!userId) throw new HTTPException(400, { message: "Missing userId" });
 
 		const updatedLink = await this.linkRepository.updateLink({
 			id,
+			index,
 			title,
-			href,
+			url,
+			userId,
 		});
 
+		if (!updatedLink)
+			throw new HTTPException(404, {
+				message: "Link not found or access denied",
+			});
+
 		return updatedLink;
-	}
+	};
 
-	async deleteLink({ id }: deleteLinkType, userId: string) {
-		const link = await this.linkRepository.getLinkById({ id });
+	deleteLink = async ({ id, userId }: deleteLinkType) => {
+		if (!id) throw new HTTPException(400, { message: "Missing id" });
+		if (!userId) throw new HTTPException(400, { message: "Missing userId" });
 
-		if (!link) {
-			throw new HTTPException(404, { message: "Link not found" });
-		}
+		const deletedLink = await this.linkRepository.deleteLink({ id, userId });
 
-		if (link.userId !== userId) {
-			throw new HTTPException(403, { message: "Forbidden" });
-		}
-
-		const deletedLink = await this.linkRepository.deleteLink({ id });
+		if (!deletedLink)
+			throw new HTTPException(404, {
+				message: "Link not found or access denied",
+			});
 
 		return deletedLink;
-	}
+	};
 }
